Wave 1 (A-E) and Wave 2 (A-B) fixes
Wave 1: Critical Bug Fixes
- A1: Fixed 2 failing negative tests (scope assertions)
- B1: Fixed UTF-8 panic in InputField (delete/backspace/move operations)
- B3: Fixed / key binding conflict in TUI runner
- C1: Fixed PortScanResults::ports_scanned overflow (u16 -> u32)
- C2: Fixed grammar fuzzer payload type mislabeling (GrammarKind enum)
- C3: Fixed CSV export unescaped fields (severity, cve_ids)
- C4: Fixed ConfigError::Io losing error chain (String -> std::io::Error)
- C5: Fixed SlapperConfig::load/save signatures (PathBuf -> impl AsRef<Path>)
- D1: Fixed infinite hang in run_packet_capture (added timeout)
- D2: Fixed unreachable arms in cycle_export_format/get_export_extension
- D3: Fixed get_tab_status unused in Resume tab
- D4: Fixed pending_key not cleared on mode change
- E2: Fixed unreachable None branch in WafEngine::run

Wave 2: Security & Error Handling
- A1: Redacted secrets in HTTP options popup (proxy_auth, auth, bearer, cookie, api_key)
- B1: Added From<tokio::time::error::Elapsed> for SlapperError
- B2: Fixed check_scope to use SlapperError (with ScopeError conversion)
- B3: Removed unused StreamEvent::error function
- B4: Log tool registration failures instead of silently ignoring
- B5: Fixed SlapperError::with_timeout unnecessary mem::take

Deferred:
- E1: Duplicate subdomain enumeration in takeover check (requires complex restructuring)

Tests: 559 passing

+++ crates/slapper/src/scanner/ports/mod.rs
@@ -66,7 +66,7 @@ pub struct PortResult {
 #[derive(Debug, Serialize, Deserialize)]
 pub struct PortScanResults {
     pub host: String,
-    pub ports_scanned: u16,
+    pub ports_scanned: u32,
     pub open_ports: Vec<PortResult>,
     pub duration_ms: u64,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -346,7 +346,7 @@ pub async fn scan_ports(
 
     Ok(PortScanResults {
         host: host.to_string(),
-        ports_scanned: ports_count as u16,
+        ports_scanned: ports_count as u32,
         open_ports: results,
         duration_ms: start.elapsed().as_millis() as u64,
         spoof_stats: None,
